# Api Player(VideosLCIII)/docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'proyectoBasico'","description":"'Demo project for Spring Boot'","contact":{"name":"Gerardo Valdez","email":"ger4rdovaldez@gmail.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'Demo project for Spring Boot'"}],"paths":{"/players":{"post":{"tags":["player-controller"],"summary":"Crea un nuevo jugador","description":"Crea un nuevo jugador con la información proporcionada en el cuerpo de la solicitud.","operationId":"createdPlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Player"}}},"required":true},"responses":{"400":{"description":"Solicitud inválida","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Operación exitosa","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"500":{"description":"Error interno del servidor","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/players/{id}":{"get":{"tags":["player-controller"],"summary":"Obtiene un jugador por su ID","description":"Devuelve un jugador basado en el ID proporcionado.","operationId":"getPlayerById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Jugador no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Operación exitosa","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"500":{"description":"Error interno del servidor","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/ping":{"get":{"tags":["health-app-controller"],"summary":"Chequea la salud de la app.","description":"Si la app esta bien , responde pong.","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}}},"components":{"schemas":{"Player":{"required":["email","password","userName"],"type":"object","properties":{"id":{"title":"Player ID","type":"integer","description":"El ID del jugador","format":"int64","example":1},"userName":{"title":"Nombre de jugador","type":"string","description":"El nombre del jugador","example":"MiNombre"},"password":{"title":"Contraseña de jugador","type":"string","description":"La contraseña del jugador","example":"MiContraseña"},"email":{"title":"Email de jugador","type":"string","description":"El Email del jugador","example":"email@email.com"},"avatar":{"title":"Avatar de jugador","type":"string","description":"El avatar del jugador","nullable":true,"example":"www.avatars.com/23"},"lastLoginDate":{"title":"Última conexión de jugador","type":"string","description":"La ultima conexión del jugador","format":"date-time","nullable":true}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}}}}}